requirements.txt
msal
Office365-REST-Python-Client
msgraph-sdk
azure-identity